requirements.txt
Django==1.10.4
django-crispy-forms==1.6.1
django-registration-redux==1.4
django-summernote==0.8.6
html2text==2014.12.29
Pillow==3.4.2
six==1.10.0
